# [now-next] Optimize lambda creation