--- a/.github/workflows/publish.yml
+++ b/.github/workflows/publish.yml
@@ -24,7 +24,7 @@ jobs:
       with:
         node-version: ${{ matrix.node-version }}
     - name: Install
-      run: yarn install --pure-lockfile
+      run: yarn install --check-files --frozen-lockfile
     - name: Build
       run: yarn build
     - name: Publish