--- a/packages/now-next/global.d.ts
+++ b/packages/now-next/global.d.ts
@@ -1,0 +1,2 @@
+// needed for @types/JSZip since we aren't building in browser mode
+interface Blob {}