--- a/packages/now-next/package.json
+++ b/packages/now-next/package.json
@@ -22,6 +22,7 @@
     "@types/resolve-from": "5.0.1",
     "@types/semver": "6.0.0",
     "@zeit/node-file-trace": "0.2.13",
+    "async-sema": "3.0.1",
     "fs-extra": "7.0.0",
     "get-port": "5.0.0",
     "resolve-from": "5.0.0",