--- a/packages/now-next/package.json
+++ b/packages/now-next/package.json
@@ -18,6 +18,7 @@
     "dist"
   ],
   "devDependencies": {
+    "@types/jszip": "3.1.6",
     "@types/next-server": "8.0.0",
     "@types/resolve-from": "5.0.1",
     "@types/semver": "6.0.0",