--- a/packages/now-next/tsconfig.json
+++ b/packages/now-next/tsconfig.json
@@ -9,6 +9,6 @@
     "sourceMap": false,
     "declaration": false
   },
-  "include": ["src/**/*", "global.d.ts"],
+  "include": ["src/**/*", "types/**/*"],
   "exclude": ["node_modules"]
 }