--- a/packages/now-node-bridge/package.json
+++ b/packages/now-node-bridge/package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "@types/aws-lambda": "8.10.19",
+    "@types/node": "10.x",
     "typescript": "3.5.2"
   }
 }