--- a/packages/now-node-bridge/src/bridge.ts
+++ b/packages/now-node-bridge/src/bridge.ts
@@ -1,3 +1,4 @@
+/// <reference types="node" />
 import { AddressInfo } from 'net';
 import { APIGatewayProxyEvent, Context } from 'aws-lambda';
 import {