--- a/packages/now-node/package.json
+++ b/packages/now-node/package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@babel/core": "7.5.0",
     "@babel/plugin-transform-modules-commonjs": "7.5.0",
+    "@types/aws-lambda": "8.10.19",
     "@types/content-type": "1.1.3",
     "@types/cookie": "0.3.3",
     "@types/etag": "1.8.0",