--- a/packages/now-ruby/package.json
+++ b/packages/now-ruby/package.json
@@ -20,7 +20,8 @@
     "prepublishOnly": "./build.sh"
   },
   "devDependencies": {
-    "execa": "^1.0.0",
+    "@types/fs-extra": "8.0.0",
+    "execa": "2.0.4",
     "fs-extra": "^7.0.1",
     "typescript": "3.5.2"
   }