--- a/packages/now-ruby/tsconfig.json
+++ b/packages/now-ruby/tsconfig.json
@@ -11,7 +11,6 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "outDir": "dist",
-    "types": ["node"],
     "strict": true,
     "target": "esnext"
   }