--- a/packages/now-static-build/package.json
+++ b/packages/now-static-build/package.json
@@ -18,6 +18,7 @@
     "prepublishOnly": "./build.sh"
   },
   "devDependencies": {
+    "@types/cross-spawn": "6.0.0",
     "@types/promise-timeout": "1.3.0",
     "cross-spawn": "6.0.5",
     "get-port": "5.0.0",