This updates the zipping of lambdas to create seed buffers with common files already compressed so that they don't need to be re-compressed for each lambda. This decreases lambda creation times by approximately **~96.23%~** **60%** (had to replace `yazl` with `jszip` due to `yazl` not handling the `forceZip64Format` option correctly

<details>
<summary>Pre-optimize Minimal</summary>

```sh
Creating lambda for page: "_error.js"...: 198.406ms
Creating lambda for page: "another copy 10.js"...: 372.232ms
Creating lambda for page: "another copy 11.js"...: 517.071ms
Creating lambda for page: "another copy 12.js"...: 689.356ms
Creating lambda for page: "another copy 2.js"...: 850.774ms
Creating lambda for page: "another copy 3.js"...: 990.463ms
Creating lambda for page: "another copy 4.js"...: 1149.230ms
Creating lambda for page: "another copy 5.js"...: 1294.805ms
Creating lambda for page: "another copy 6.js"...: 1456.029ms
Creating lambda for page: "another copy 7.js"...: 1594.930ms
Creating lambda for page: "another copy 8.js"...: 1752.720ms
Creating lambda for page: "another copy 9.js"...: 1913.041ms
Creating lambda for page: "another copy.js"...: 2067.628ms
Creating lambda for page: "another.js"...: 2196.671ms
```
</details>

<details>
<summary>Post-optimize Minimal</summary>

```sh
Creating Zipped Seed Buffers: 164.155ms
Creating lambda for page: "_error.js"...: 98.827ms
Creating lambda for page: "another copy 12.js"...: 108.911ms
Creating lambda for page: "another copy 11.js"...: 108.995ms
Creating lambda for page: "another copy 10.js"...: 109.072ms
Creating lambda for page: "another copy 2.js"...: 123.629ms
Creating lambda for page: "another copy 6.js"...: 123.652ms
Creating lambda for page: "another copy 4.js"...: 140.528ms
Creating lambda for page: "another copy 3.js"...: 140.626ms
Creating lambda for page: "another copy 5.js"...: 140.644ms
Creating lambda for page: "another copy 7.js"...: 140.648ms
Creating lambda for page: "another.js"...: 168.109ms
Creating lambda for page: "another copy 9.js"...: 169.200ms
Creating lambda for page: "another copy.js"...: 170.785ms
Creating lambda for page: "another copy 8.js"...: 172.387ms
```
</details>